Add support to load invalid file paths
Add support to load invalid file paths so that if file not exists, a new
instance of InvalidRUbioSeqFile is created (instead of returning null
which was the previous behaviour).

Within this change, invalid paths are represented by InvalidRUbioSeqFile
objects and its GUI component sclass property can be established to
highlight that the file is not valid.

It is important to note that in the MethylationExperiment the
projectCompletePath may not exist in the moment of editing the
experiment so we have to test only if the parent of the
projectCompletePath exists.

This commit also introduces a new feature in the Execute experiment
screen. Now, every time a new XML configuration file is selected its
content is loaded in order to validate the parameters (invalid paths
above all). If an invalid configuration is loaded, a warning is showed
to the user and the file is not selected, avoiding running RUbioSeq with
invalid XML files.