Refactor computation of the number of stages of an experiment
Refactor computation of the number of stages of an experiment so that
Measurable interface declares a new method getStagesCount(int
executionLevel) that the four types of experiments must be implement.
This change allows to move the responsibility of computing the number of
stages from the ViewExperimentModel to the concrete types of experiment. 

To be sure that this refactor worked well a new text
(ExperimentStagesTest) was added before doing any change. Initially,
this class tested the initialiceStagesCount function of the
ViewExperimentModel. Once all cases were covered, the previous refactor
was applied and then modified the test to check that all cases were
being calculated right.